# api/openapi.yaml
openapi: 3.1.0
info:
  description: ""
  license:
    name: API Pagos Instantáneos
    url: https://khipu.com
  title: API Pagos Instantáneos
  version: v3.0
servers:
- description: Producción
  url: https://payment-api.khipu.com
security:
- Api-Key: []
paths:
  /v3/banks:
    get:
      description: Este método obtiene la lista de bancos que se pueden utilizar para
        pagar en esta cuenta de cobro.
      operationId: getBanks
      parameters: []
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/bank-get-banks'
          description: Operación exitosa. Devuelve un objeto JSON con el conjunto
            de bancos disponibles.
        "400":
          description: Bad request - Solicitud mal formada
        "401":
          description: Unauthorized - Credenciales no válidas o faltantes
        "403":
          description: Forbidden - Cliente no tiene privilegios
        "404":
          description: Not found - Recurso no disponible
        "500":
          description: Internal server error - Servidor no pudo completar la solicitud
      summary: Get banks
  /v3/payments/{id}/confirm:
    post:
      description: |
        **Advertencia: Esta función sólo está disponible para los clientes que la hayan contratado de forma independiente.
        Para utilizarla, póngase en contacto con nosotros en soporte@khipu.com**<br><br>
        Confirmar el pago. Al confirmar el pago, este será rendido al día hábil
        siguiente.
      operationId: postPaymentConfirmById
      parameters:
      - description: Identificador del pago
        explode: false
        in: path
        name: id
        required: true
        schema:
          maxLength: 255
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/success'
          description: Operación exitosa.
        "400":
          description: Bad request - Solicitud mal formada
        "401":
          description: Unauthorized - Credenciales no válidas o faltantes
        "403":
          description: Forbidden - Cliente no tiene privilegios
        "404":
          description: Not found - Recurso no disponible
        "500":
          description: Internal server error - Servidor no pudo completar la solicitud
      summary: Confirm payment by Id
  /v3/payments/{id}:
    delete:
      description: Borrar un pago. Solo se pueden borrar pagos que estén pendientes
        de pagar. Esta operación no puede deshacerse.
      operationId: deletePaymentById
      parameters:
      - description: Identificador del pago
        explode: false
        in: path
        name: id
        required: true
        schema:
          maxLength: 255
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/success'
          description: Operación exitosa.
        "400":
          description: Bad request - Solicitud mal formada
        "401":
          description: Unauthorized - Credenciales no válidas o faltantes
        "403":
          description: Forbidden - Cliente no tiene privilegios
        "404":
          description: Not found - Recurso no disponible
        "500":
          description: Internal server error - Servidor no pudo completar la solicitud
      summary: Delete payment by Id
    get:
      description: Información completa del pago. Datos con los que fue creado y el
        estado actual del pago.
      operationId: getPaymentById
      parameters:
      - description: Identificador del pago
        explode: false
        in: path
        name: id
        required: true
        schema:
          maxLength: 255
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/payment-get-payment'
          description: Operación exitosa. Devuelve un objeto JSON con la información
            de pago.
        "400":
          description: Bad request - Solicitud mal formada
        "401":
          description: Unauthorized - Credenciales no válidas o faltantes
        "403":
          description: Forbidden - Cliente no tiene privilegios
        "404":
          description: Not found - Recurso no disponible
        "500":
          description: Internal server error - Servidor no pudo completar la solicitud
      summary: Get payment by Id
  /v3/payments:
    post:
      description: Crea un pago en Khipu y obtiene las URLs para redirección al usuario
        para que complete el pago.
      operationId: postPayment
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/payment-post-payment'
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/payment-post-payment-2'
          description: Operación exitosa. Devuelve un objeto JSON con la información
            del pago creado.
        "400":
          description: Bad request - Solicitud mal formada
        "401":
          description: Unauthorized - Credenciales no válidas o faltantes
        "403":
          description: Forbidden - Cliente no tiene privilegios
        "404":
          description: Not found - Recurso no disponible
        "500":
          description: Internal server error - Servidor no pudo completar la solicitud
      summary: Create payment
  /v3/payments/{id}/refunds:
    post:
      description: Reembolsa total o parcialmente el monto de un pago. Esta operación
        solo se puede realizar en los comercios que recauden en cuenta Khipu y antes
        de la rendición de los fondos correspondientes.
      operationId: postPaymentRefundsById
      parameters:
      - description: Identificador del pago
        explode: false
        in: path
        name: id
        required: true
        schema:
          maxLength: 255
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/payment-post-payment-refunds'
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/success'
          description: Operación exitosa.
        "400":
          description: Bad request - Solicitud mal formada
        "401":
          description: Unauthorized - Credenciales no válidas o faltantes
        "403":
          description: Forbidden - Cliente no tiene privilegios
        "404":
          description: Not found - Recurso no disponible
        "500":
          description: Internal server error - Servidor no pudo completar la solicitud
      summary: Refund payment by Id
  /v3/predict:
    get:
      description: "Predicción acerca del resultado de un pago, si podrá o no funcionar.\
        \ Información adicional como máximo posible de transferir a un nuevo destinatario."
      operationId: getPredict
      parameters:
      - description: Correo electrónico del pagador
        explode: true
        in: query
        name: payer_email
        required: true
        schema:
          maxLength: 255
          type: string
        style: form
      - description: Identificador del banco de origen
        explode: true
        in: query
        name: bank_id
        required: true
        schema:
          maxLength: 255
          type: string
        style: form
      - description: Monto del pago
        explode: true
        in: query
        name: amount
        required: true
        schema:
          maxLength: 255
          type: string
        style: form
      - description: Moneda en formato ISO-4217
        explode: true
        in: query
        name: currency
        required: true
        schema:
          maxLength: 255
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/predict-get-predict'
          description: Operación exitosa. Devuelve un objeto JSON con la predicción
            junto con información adiciónal.
        "400":
          description: Bad request - Solicitud mal formada
        "401":
          description: Unauthorized - Credenciales no válidas o faltantes
        "403":
          description: Forbidden - Cliente no tiene privilegios
        "404":
          description: Not found - Recurso no disponible
        "500":
          description: Internal server error - Servidor no pudo completar la solicitud
      summary: Get payment prediction
  /v3/receivers:
    post:
      description: |
        **Advertencia: Esta función sólo está disponible para los clientes que la hayan contratado de forma independiente.
        Para utilizarla, póngase en contacto con nosotros en soporte@khipu.com**<br><br>
        Crear una nueva cuenta de cobro asociada a un integrador. Necesita datos
        de la cuenta de usuario asociada, datos de facturación y datos de contacto.
      operationId: postReceiver
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/receiver-post-receiver'
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/receiver-post-receiver-2'
          description: Operación exitosa. Devuelve un objeto JSON con el identificador
            y secreto de la nueva cuenta de cobro.
        "400":
          description: Bad request - Solicitud mal formada
        "401":
          description: Unauthorized - Credenciales no válidas o faltantes
        "403":
          description: Forbidden - Cliente no tiene privilegios
        "404":
          description: Not found - Recurso no disponible
        "500":
          description: Internal server error - Servidor no pudo completar la solicitud
      summary: Post receiver
  /v3/merchants/{id}/paymentMethods:
    get:
      description: Obtiene el listado de medios de pago disponible para una cuenta
        de cobrador.
      operationId: getMerchantPaymentMethodsById
      parameters:
      - description: Identificador de la cuenta de cobro
        explode: false
        in: path
        name: id
        required: true
        schema:
          type: number
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/merchant-get-payment-methods'
          description: Operación exitosa. Devuelve un objeto JSON con la variedad
            de métodos de pago disponibles.
        "400":
          description: Bad request - Solicitud mal formada
        "401":
          description: Unauthorized - Credenciales no válidas o faltantes
        "403":
          description: Forbidden - Cliente no tiene privilegios
        "404":
          description: Not found - Recurso no disponible
        "500":
          description: Internal server error - Servidor no pudo completar la solicitud
      summary: Get payment methods
components:
  schemas:
    bank-get-banks:
      example:
        banks:
        - parent: ""
          logo_url: https://s3.amazonaws.com/static.khipu.com/logos/bancos/chile/estado-icon.png
          bank_id: SDdGj
          min_amount: 1000
          name: Banco Estado
          message: "Tarifa de $300 de transferencia a otros bancos, usando CuentaRUT."
          type: Persona
        - parent: ""
          logo_url: https://s3.amazonaws.com/static.khipu.com/logos/bancos/chile/estado-icon.png
          bank_id: SDdGj
          min_amount: 1000
          name: Banco Estado
          message: "Tarifa de $300 de transferencia a otros bancos, usando CuentaRUT."
          type: Persona
      properties:
        banks:
          description: Arreglo con listado de bancos.
          items:
            $ref: '#/components/schemas/bank_get_banks_banks_inner'
          minItems: 1
          type: array
      required:
      - banks
    success:
      description: Objeto para respuestas exitosas de la API.
      example:
        message: Message.
      properties:
        message:
          description: Mensaje a desplegar al usuario.
          example: Message.
          type: string
      required:
      - message
    payment-get-payment:
      description: Información completa del pago. Datos con los que fue creado y el
        estado actual del pago.
      example:
        conciliation_date: 2017-03-01T13:00:00.000Z
        payment_url: https://khipu.com/payment/info/gqzdy6chjne9
        ready_for_terminal: false
        bank_account_number: "001120490689"
        picture_url: https://micomercio.com/picture_url
        payer_email: pagador@email.com
        responsible_user_email: responsible@email.com
        receiver_id: 985101
        subject: Test
        discount: 0
        send_reminders: true
        body: Test
        bank: Banco de Chile (Edwards Citi)
        receipt_url: https://micomercio.com/order/receipt_url
        funds_source: debit
        payment_id: gqzdy6chjne9
        bank_id: dfFbF
        return_url: https://micomercio.com/order/return_url
        transfer_url: https://khipu.com/payment/manual/gqzdy6chjne9
        currency: CLP
        payment_method: simplified_transfer
        transaction_id: zwo3wqz6uulcvajt
        amount: 1000
        app_url: khipu:///pos/gqzdy6chjne9
        attachment_urls:
        - https://micomercio.com/attachment1.pdf
        - https://micomercio.com/attachment1.pdf
        out_of_date_conciliation: true
        notification_token: 9dec8aa176c5223026919b3b5579a4776923e646ff3be686b9e6b62ec042e91f
        expires_date: 2023-12-31T15:45:00-04:00
        custom: <xml>...</xml>
        send_email: true
        notify_url: https://micomercio.com/webhook/notify_url
        status_detail: normal
        third_party_authorization_details: third_party_authorization_details
        notify_api_version: "3.0"
        personal_identifier: 11.000.111-9
        payer_name: Nombre Pagador
        cancel_url: https://micomercio.com/order/cancel_url
        simplified_transfer_url: https://app.khipu.com/payment/simplified/gqzdy6chjne9
        status: done
      properties:
        payment_id:
          description: "Identificador único del pago, es una cadena alfanumérica de\
            \ 12 caracteres. Como este identificador es único, se puede usar, por\
            \ ejemplo, para evitar procesar una notificación repetida. (Khipu espera\
            \ un código 200 al notificar un pago, si esto no ocurre se reintenta hasta\
            \ por dos días)."
          example: gqzdy6chjne9
          type: string
        payment_url:
          description: "URL principal del pago, si el usuario no ha elegido previamente\
            \ un método de pago se le muestran las opciones."
          example: https://khipu.com/payment/info/gqzdy6chjne9
          type: string
        simplified_transfer_url:
          description: URL de pago simplificado.
          example: https://app.khipu.com/payment/simplified/gqzdy6chjne9
          type: string
        transfer_url:
          description: URL de pago normal.
          example: https://khipu.com/payment/manual/gqzdy6chjne9
          type: string
        app_url:
          description: URL para invocar el pago desde un dispositivo móvil usando
            la APP de Khipu.
          example: khipu:///pos/gqzdy6chjne9
          type: string
        ready_for_terminal:
          description: Es `true` si el pago ya cuenta con todos los datos necesarios
            para abrir directamente la aplicación de pagos Khipu.
          example: false
          type: boolean
        notification_token:
          description: "Cadena de caracteres alfanuméricos que identifican unicamente\
            \ al pago, es el identificador que el servidor de Khipu enviará al servidor\
            \ del comercio cuando notifique que un pago está conciliado."
          example: 9dec8aa176c5223026919b3b5579a4776923e646ff3be686b9e6b62ec042e91f
          type: string
        receiver_id:
          description: Identificador único de una cuenta de cobro.
          example: 985101
          format: int64
          type: integer
        conciliation_date:
          description: Fecha y hora de conciliación del pago. Formato ISO-8601.
          example: 2017-03-01T13:00:00.000Z
          format: date-time
          type: string
        subject:
          description: Motivo del pago.
          example: Test
          type: string
        amount:
          description: El monto del cobro.
          example: 1000
          format: double
          type: number
        currency:
          description: El código de moneda en formato ISO-4217.
          example: CLP
          type: string
        status:
          description: "Estado del pago, puede ser `pending` (el pagador aún no comienza\
            \ a pagar), `verifying` (se está verificando el pago) o `done`, cuando\
            \ el pago ya está confirmado."
          enum:
          - pending
          - verifying
          - done
          example: done
          type: string
        status_detail:
          description: |-
            Detalle del estado del pago:
            `pending` (el pagador aún no comienza a pagar),
            `normal` (el pago fue verificado y fue cancelado por algún medio de pago estándar),
            `marked-paid-by-receiver` (el cobrador marcó el cobro como pagado por otro medio),
            `rejected-by-payer` (el pagador declaró que no pagará),
            `marked-as-abuse` (el pagador declaró que no pagará y que el cobro fue no solicitado), y
            `reversed` (el pago fue anulado por el comercio, el dinero fue devuelto al pagador).
          enum:
          - pending
          - normal
          - marked-paid-by-receiver
          - rejected-by-payer
          - marked-as-abuse
          - reversed
          example: normal
          type: string
        body:
          description: Detalle del cobro.
          example: Test
          type: string
        picture_url:
          description: URL con imagen del cobro.
          example: https://micomercio.com/picture_url
          type: string
        receipt_url:
          description: URL del comprobante de pago.
          example: https://micomercio.com/order/receipt_url
          type: string
        return_url:
          description: URL donde se redirige al pagador luego que termina el pago.
          example: https://micomercio.com/order/return_url
          type: string
        cancel_url:
          description: URL donde se redirige al pagador luego de que desiste hacer
            el pago.
          example: https://micomercio.com/order/cancel_url
          type: string
        notify_url:
          description: URL del webservice donde se notificará el pago.
          example: https://micomercio.com/webhook/notify_url
          type: string
        notify_api_version:
          description: Versión de la API de notificación.
          example: "3.0"
          type: string
        expires_date:
          description: Fecha máxima para ejecutar el pago (en formato ISO-8601). El
            cliente podrá realizar varios intentos de pago hasta dicha fecha. Cada
            intento tiene un plazo individual de 3 horas para su ejecución.
          example: 2023-12-31T15:45:00-04:00
          format: date-time
          type: string
        attachment_urls:
          description: Arreglo de URLs de archivos adjuntos al pago.
          items:
            description: URL de archivo adjunto al pago.
            example: https://micomercio.com/attachment1.pdf
            type: string
          minItems: 0
          type: array
        bank:
          description: Nombre del banco seleccionado por el pagador.
          example: Banco de Chile (Edwards Citi)
          type: string
        bank_id:
          description: Identificador del banco seleccionado por el pagador.
          example: dfFbF
          type: string
        payer_name:
          description: Nombre del pagador.
          example: Nombre Pagador
          type: string
        payer_email:
          description: Correo electrónico del pagador.
          example: pagador@email.com
          format: email
          type: string
        personal_identifier:
          description: Identificador personal del pagador.
          example: 11.000.111-9
          type: string
        bank_account_number:
          description: Número de cuenta bancaria del pagador.
          example: "001120490689"
          type: string
        out_of_date_conciliation:
          description: Es `true` si la conciliación del pago fue hecha luego de la
            fecha de expiración.
          example: true
          type: boolean
        transaction_id:
          description: Identificador del pago asignado por el cobrador.
          example: zwo3wqz6uulcvajt
          type: string
        custom:
          description: Campo genérico que asigna el cobrador al momento de hacer el
            pago.
          example: <xml>...</xml>
          type: string
        responsible_user_email:
          description: Correo electrónico de la persona responsable del pago.
          example: responsible@email.com
          type: string
        send_reminders:
          description: Es `true` cuando este es un cobro por correo electrónico y
            Khipu enviará recordatorios.
          example: true
          type: boolean
        send_email:
          description: Es `true` cuando Khipu enviará el cobro por correo electrónico.
          example: true
          type: boolean
        payment_method:
          description: "Método de pago usado por el pagador, puede ser `regular_transfer`\
            \ (transferencia normal) o `simplified_transfer` (transferencia simplificada)."
          enum:
          - regular_transfer
          - simplified_transfer
          - not_available
          example: simplified_transfer
          type: string
        funds_source:
          description: "Origen de fondos usado por el pagador, puede ser `debit` para\
            \ pago con débito, `prepaid` para pago con prepago, `credit` para pago\
            \ con crédito, o vacío en el caso de que se haya pagado mediante transferencia\
            \ bancaria."
          enum:
          - debit
          - prepaid
          - credit
          - not-available
          - ""
          example: debit
          type: string
        discount:
          description: Monto a descontar del valor pagado.
          example: 0
          format: double
          type: number
        third_party_authorization_details:
          description: Ignorar este campo.
          type: string
      required:
      - amount
      - app_url
      - attachment_urls
      - bank
      - bank_account_number
      - bank_id
      - body
      - cancel_url
      - conciliation_date
      - currency
      - custom
      - expires_date
      - funds_source
      - notification_token
      - notify_api_version
      - notify_url
      - out_of_date_conciliation
      - payer_email
      - payer_name
      - payment_id
      - payment_method
      - payment_url
      - personal_identifier
      - picture_url
      - ready_for_terminal
      - receipt_url
      - receiver_id
      - responsible_user_email
      - return_url
      - send_email
      - send_reminders
      - simplified_transfer_url
      - status
      - status_detail
      - subject
      - transaction_id
      - transfer_url
...
    payment-post-payment-2:
      description: Crea un pago en Khipu y obtiene las URLs para redirección al usuario
        para que complete el pago.
      example:
        payment_url: https://khipu.com/payment/info/gqzdy6chjne9
        ready_for_terminal: false
        app_url: khipu:///pos/gqzdy6chjne9
        payment_id: gqzdy6chjne9
        transfer_url: https://khipu.com/payment/manual/gqzdy6chjne9
        simplified_transfer_url: https://app.khipu.com/payment/simplified/gqzdy6chjne9
      properties:
        payment_id:
          description: "Identificador único del pago, es una cadena alfanumérica de\
            \ 12 caracteres. Como este identificador es único, se puede usar, por\
            \ ejemplo, para evitar procesar una notificación repetida. (Khipu espera\
            \ un código 200 al notificar un pago, si esto no ocurre se reintenta hasta\
            \ por dos días)."
          example: gqzdy6chjne9
          type: string
        payment_url:
          description: "URL principal del pago, si el usuario no ha elegido previamente\
            \ un método de pago se le muestran las opciones."
          example: https://khipu.com/payment/info/gqzdy6chjne9
          type: string
        simplified_transfer_url:
          description: URL de pago simplificado.
          example: https://app.khipu.com/payment/simplified/gqzdy6chjne9
          type: string
        transfer_url:
          description: URL de pago normal.
          example: https://khipu.com/payment/manual/gqzdy6chjne9
          type: string
        app_url:
          description: URL para invocar el pago desde un dispositivo móvil usando
            la APP de Khipu.
          example: khipu:///pos/gqzdy6chjne9
          type: string
        ready_for_terminal:
          description: Es `true` si el pago ya cuenta con todos los datos necesarios
            para abrir directamente la aplicación de pagos Khipu.
          example: false
          type: boolean
      required:
      - app_url
      - payment_id
      - payment_url
      - ready_for_terminal
      - simplified_transfer_url
      - transfer_url
    payment-post-payment-refunds:
      description: Reembolsa total o parcialmente el monto de un pago. Esta operación
        solo se puede realizar en los comercios que recauden en cuenta Khipu y antes
        de la rendición de los fondos correspondientes.
      properties:
        amount:
          description: "El monto a devolver. Sin separador de miles y usando '.' como\
            \ separador de decimales. Hasta 4 lugares decimales, dependiendo de la\
            \ moneda. Si se omite el reembolso se hará por el total del monto del\
            \ pago."
          example: 1000
          format: double
          type: number
...
  securitySchemes:
    Api-Key:
      in: header
      name: x-api-key
      type: apiKey